backend/requirements.txt
crewai
langchain-anthropic
anthropic
httpx
yfinance
pandas
numpy
supabase
feedparser
beautifulsoup4
apscheduler
fastapi
uvicorn
python-dotenv
loguru
redis
google-generativeai

pypdf
pymupdf
